requirements.txt
fastapi>=0.110.0
uvicorn[standard]
python-multipart
openai>=1.0.0
pydantic>=2.0.0
pydantic-settings
pymupdf
pymongo>=4.0.0
pyyaml>=6.0
